Refactor backend into routers
